+++ CoreMapView/src/au/gov/asd/tac/constellation/views/mapview/layers/ActivityHeatmapLayer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2021 Australian Signals Directorate
+ * Copyright 2010-2024 Australian Signals Directorate
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -42,14 +42,10 @@ public class ActivityHeatmapLayer extends AbstractHeatmapLayer {
             try {
                 for (final GraphElement element : renderer.getMarkerCache().get(marker)) {
                     switch (element.getType()) {
-                        case VERTEX:
-                            activityCount += readableGraph.getVertexTransactionCount(element.getId());
-                            break;
-                        case TRANSACTION:
-                            activityCount += 1;
-                            break;
-                        default:
-                            break;
+                        case VERTEX -> activityCount += readableGraph.getVertexTransactionCount(element.getId());
+                        case TRANSACTION -> activityCount += 1;
+                        default -> {
+                        }
                     }
                 }
             } finally {