+++ CoreMapView/src/au/gov/asd/tac/constellation/views/mapview/markers/ConstellationAbstractMarker.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2021 Australian Signals Directorate
+ * Copyright 2010-2024 Australian Signals Directorate
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -183,8 +183,8 @@ public abstract class ConstellationAbstractMarker implements Marker {
     @Override
     public String getStringProperty(final String key) {
         final Object value = properties.get(key);
-        if (value instanceof String) {
-            return (String) value;
+        if (value instanceof String string) {
+            return string;
         } else {
             return null;
         }
@@ -193,8 +193,8 @@ public abstract class ConstellationAbstractMarker implements Marker {
     @Override
     public Integer getIntegerProperty(final String key) {
         final Object value = properties.get(key);
-        if (value instanceof Integer) {
-            return (Integer) value;
+        if (value instanceof Integer integer) {
+            return integer;
         } else {
             return null;
         }