+++ CoreUtilities/src/au/gov/asd/tac/constellation/utilities/datastructure/ObjectCache.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2021 Australian Signals Directorate
+ * Copyright 2010-2024 Australian Signals Directorate
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -85,7 +85,7 @@ public class ObjectCache<K extends Object, V extends Object> {
             try {
                 key = entry.getKey();
                 value = entry.getValue();
-            } catch (IllegalStateException ex) {
+            } catch (final IllegalStateException ex) {
                 throw new ConcurrentModificationException(ex);
             }
             action.accept(key, value);