+++ CoreWordCloudView/test/unit/src/au/gov/asd/tac/constellation/views/wordcloud/ui/WordCloudPaneNGTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2023 Australian Signals Directorate
+ * Copyright 2010-2024 Australian Signals Directorate
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -32,6 +32,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
 import org.testfx.api.FxToolkit;
+import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 import org.testng.annotations.AfterClass;
@@ -42,7 +43,7 @@ import org.testng.annotations.Test;
 
 /**
  * Test class for WordCloudPane
- * 
+ *
  * @author Delphinus8821
  */
 public class WordCloudPaneNGTest {
@@ -85,7 +86,7 @@ public class WordCloudPaneNGTest {
             controllerStatic.when(WordCloudController::getDefault).thenReturn(controller);
             when(controller.init(topComponent)).thenReturn(controller);
             final WordCloudPane instance = new WordCloudPane(controller);
-            
+
             instance.setInProgress();
             assertTrue(instance.getCloudStackPane().getChildren().contains(instance.getSpinner()));
         }
@@ -122,16 +123,16 @@ public class WordCloudPaneNGTest {
             boolean unionButtonSelected = true;
             boolean frequencyButtonSelected = false;
             boolean hasSignificances = false;
-            instance.getTheCloud().getChildren().add(instance.getSliderBar());
-            
+            instance.getTheCloud().getChildren().add(instance.getSliderBarPane());
+
             instance.enableTheCloud(unionButtonSelected, frequencyButtonSelected, hasSignificances);
-            assertFalse(instance.getTheCloud().getChildren().contains(instance.getSliderBar()));
-            
+            assertFalse(instance.getTheCloud().getChildren().contains(instance.getSliderBarPane()));
+
             unionButtonSelected = false;
             frequencyButtonSelected = true;
             hasSignificances = true;
             instance.enableTheCloud(unionButtonSelected, frequencyButtonSelected, hasSignificances);
-            assertTrue(instance.getTheCloud().getChildren().contains(instance.getSliderBar()));
+            assertTrue(instance.getTheCloud().getChildren().contains(instance.getSliderBarPane()));
         }
     }
 
@@ -170,7 +171,7 @@ public class WordCloudPaneNGTest {
             final String queryInfo = "Phrase length 1";
             final int baseFontSize = 12;
             instance.createWords(wordListWithSizes, queryInfo, baseFontSize);
-            
+
             final Map<String, Hyperlink> result = instance.getWordButtons();
             assertTrue(result.size() == 4);
         }
@@ -187,7 +188,7 @@ public class WordCloudPaneNGTest {
             controllerStatic.when(WordCloudController::getDefault).thenReturn(controller);
             when(controller.init(topComponent)).thenReturn(controller);
             final WordCloudPane instance = new WordCloudPane(controller);
-            
+
             final SortedMap<String, Float> wordListWithSizes = new TreeMap<>();
             wordListWithSizes.put("Node", (float) 2);
             wordListWithSizes.put("Edge", (float) 4);
@@ -196,12 +197,12 @@ public class WordCloudPaneNGTest {
             final Set<String> selectedWords = new HashSet<>();
             selectedWords.add("Link");
             selectedWords.add("Node");
-            
+
             // Add words to be selected
             final String queryInfo = "Phrase length 1";
             final int baseFontSize = 12;
             instance.createWords(wordListWithSizes, queryInfo, baseFontSize);
-            
+
             // Select only 2 words
             instance.updateSelection(selectedWords);
             final Map<String, Hyperlink> result = instance.getWordButtons();
@@ -223,26 +224,43 @@ public class WordCloudPaneNGTest {
             controllerStatic.when(WordCloudController::getDefault).thenReturn(controller);
             when(controller.init(topComponent)).thenReturn(controller);
             final WordCloudPane instance = new WordCloudPane(controller);
-            
+
             // Add words to be selected
             final SortedSet<String> wordsToDisplay = new TreeSet<>();
             wordsToDisplay.add("Node");
             wordsToDisplay.add("Link");
-            
+
             // Create word buttons 
             final SortedMap<String, Float> wordListWithSizes = new TreeMap<>();
             wordListWithSizes.put("Node", (float) 2);
             wordListWithSizes.put("Edge", (float) 4);
             wordListWithSizes.put("Link", (float) 7);
-            wordListWithSizes.put("Transaction", (float) 1);    
+            wordListWithSizes.put("Transaction", (float) 1);
             final String queryInfo = "Phrase length 1";
             final int baseFontSize = 12;
             instance.createWords(wordListWithSizes, queryInfo, baseFontSize);
-            
+
             // Update words 
             final boolean reapplySort = true;
             instance.updateWords(wordsToDisplay, reapplySort);
             assertTrue(instance.getWords().getChildren().size() == 2);
+        }
+    }
+
+    /**
+     * Test of setContentHeight method, of class WordCloudPane.
+     */
+    @Test
+    public void testSetContentHeight() {
+        System.out.println("setContentHeight");
+        try (final MockedStatic<WordCloudController> controllerStatic = Mockito.mockStatic(WordCloudController.class)) {
+            final WordCloudController controller = spy(WordCloudController.class);
+            controllerStatic.when(WordCloudController::getDefault).thenReturn(controller);
+            when(controller.init(topComponent)).thenReturn(controller);
+            final WordCloudPane instance = new WordCloudPane(controller);
+            
+            instance.setContentHeight(100);
+            assertEquals(instance.getContentHeight(), 90.0d);
         }
     }
 }